tests: move create_*_file to SystemTestsMixin
